Data-Engineering/Day7/ex1/requirements.txt
flask
pandas
google-cloud-bigquery